Navigation_HomeScreen_in_separate_file_DetailScreen_added_to_test
