+++ package.json
@@ -1,5 +1,5 @@
 {
-  "private": true, 
+  "private": true,
   "name": "new-netflix",
   "version": "1.0.0",
   "description": "",
@@ -11,7 +11,7 @@
   "author": "",
   "license": "ISC",
   "workspaces": [
-      "packages/*"
+    "packages/*"
   ],
   "devDependencies": {
     "lerna": "^3.20.2"
